working; new query and scripting
