+++ src/sql/workbench/parts/query/browser/queryActions.ts
@@ -11,7 +11,6 @@ import { IContextViewService } from 'vs/platform/contextview/browser/contextView
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
 import { INotificationService } from 'vs/platform/notification/common/notification';
-import { IExtensionTipsService } from 'vs/workbench/services/extensionManagement/common/extensionManagement';
 import Severity from 'vs/base/common/severity';
 import { append, $ } from 'vs/base/browser/dom';
 
@@ -21,13 +20,27 @@ import {
 	IConnectionParams,
 	INewConnectionParams,
 	ConnectionType,
-	RunQueryOnConnectionMode
+	RunQueryOnConnectionMode,
+	IConnectionCompletionOptions,
+	IConnectableInput
 } from 'sql/platform/connection/common/connectionManagement';
 import { QueryEditor } from 'sql/workbench/parts/query/browser/queryEditor';
 import { IQueryModelService } from 'sql/platform/query/common/queryModel';
 import { SelectBox } from 'sql/base/browser/ui/selectBox/selectBox';
 import { attachEditableDropdownStyler, attachSelectBoxStyler } from 'sql/platform/theme/common/styler';
 import { Dropdown } from 'sql/base/parts/editableDropdown/browser/dropdown';
+import { Task } from 'sql/platform/tasks/browser/tasksRegistry';
+import { IObjectExplorerService } from 'sql/workbench/services/objectExplorer/common/objectExplorerService';
+import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
+import { IQueryEditorService } from 'sql/workbench/services/queryEditor/common/queryEditorService';
+import { IConnectionProfile } from 'sql/platform/connection/common/interfaces';
+import { getCurrentGlobalConnection } from 'sql/workbench/browser/taskUtilities';
+import { ServicesAccessor, IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
+import { CommandsRegistry } from 'vs/platform/commands/common/commands';
+import { OEAction } from 'sql/workbench/parts/objectExplorer/browser/objectExplorerActions';
+import { TreeViewItemHandleArg } from 'sql/workbench/common/views';
+import { ICapabilitiesService } from 'sql/platform/capabilities/common/capabilitiesService';
+import { ConnectionProfile } from 'sql/platform/connection/common/connectionProfile';
 
 /**
  * Action class that query-based Actions will extend. This base class automatically handles activating and
@@ -97,6 +110,83 @@ export abstract class QueryTaskbarAction extends Action {
 	}
 }
 
+export function openNewQuery(accessor: ServicesAccessor, profile?: IConnectionProfile, initalContent?: string, onConnection?: RunQueryOnConnectionMode): Promise<void> {
+	const editorService = accessor.get(IEditorService);
+	const queryEditorService = accessor.get(IQueryEditorService);
+	const objectExplorerService = accessor.get(IObjectExplorerService);
+	const connectionManagementService = accessor.get(IConnectionManagementService);
+	if (!profile) {
+		profile = getCurrentGlobalConnection(objectExplorerService, connectionManagementService, editorService);
+	}
+	return queryEditorService.newSqlEditor(initalContent).then((owner: IConnectableInput) => {
+		// Connect our editor to the input connection
+		let options: IConnectionCompletionOptions = {
+			params: { connectionType: ConnectionType.editor, runQueryOnCompletion: onConnection, input: owner },
+			saveTheConnection: false,
+			showDashboard: false,
+			showConnectionDialogOnError: true,
+			showFirewallRuleOnError: true
+		};
+		if (profile) {
+			return connectionManagementService.connect(profile, owner.uri, options).then();
+		}
+		return undefined;
+	});
+}
+
+// --- actions
+export class NewQueryTask extends Task {
+	public static ID = 'newQuery';
+	public static LABEL = nls.localize('newQueryTask.newQuery', "New Query");
+	public static ICON = 'new-query';
+
+	constructor() {
+		super({
+			id: NewQueryTask.ID,
+			title: NewQueryTask.LABEL,
+			iconPath: undefined,
+			iconClass: NewQueryTask.ICON
+		});
+	}
+
+	public runTask(accessor: ServicesAccessor, profile: IConnectionProfile): Promise<void> {
+		return openNewQuery(accessor, profile);
+	}
+}
+
+export const OE_NEW_QUERY_ACTION_ID = 'objectExplorer.newQuery';
+
+CommandsRegistry.registerCommand(OE_NEW_QUERY_ACTION_ID, (accessor: ServicesAccessor, actionContext: any) => {
+	const instantiationService = accessor.get(IInstantiationService);
+	return instantiationService.createInstance(OEAction, NewQueryTask.ID, NewQueryTask.LABEL).run(actionContext);
+});
+
+export const DE_NEW_QUERY_COMMAND_ID = 'dataExplorer.newQuery';
+
+// New Query
+CommandsRegistry.registerCommand({
+	id: DE_NEW_QUERY_COMMAND_ID,
+	handler: (accessor, args: TreeViewItemHandleArg) => {
+		if (args.$treeItem) {
+			const queryEditorService = accessor.get(IQueryEditorService);
+			const connectionService = accessor.get(IConnectionManagementService);
+			const capabilitiesService = accessor.get(ICapabilitiesService);
+			return queryEditorService.newSqlEditor().then((owner: IConnectableInput) => {
+				// Connect our editor to the input connection
+				let options: IConnectionCompletionOptions = {
+					params: { connectionType: ConnectionType.editor, input: owner },
+					saveTheConnection: false,
+					showDashboard: false,
+					showConnectionDialogOnError: true,
+					showFirewallRuleOnError: true
+				};
+				return connectionService.connect(new ConnectionProfile(capabilitiesService, args.$treeItem.payload), owner.uri, options);
+			});
+		}
+		return Promise.resolve(true);
+	}
+});
+
 /**
  * Action class that runs a query in the active SQL text document.
  */
@@ -108,8 +198,7 @@ export class RunQueryAction extends QueryTaskbarAction {
 	constructor(
 		editor: QueryEditor,
 		@IQueryModelService protected readonly queryModelService: IQueryModelService,
-		@IConnectionManagementService connectionManagementService: IConnectionManagementService,
-		@IExtensionTipsService private readonly extensionTipsService: IExtensionTipsService
+		@IConnectionManagementService connectionManagementService: IConnectionManagementService
 	) {
 		super(connectionManagementService, editor, RunQueryAction.ID, RunQueryAction.EnabledClass);
 		this.label = nls.localize('runQueryLabel', "Run");
@@ -424,7 +513,6 @@ export class ListDatabasesActionItem implements IActionViewItem {
 
 	public actionRunner: IActionRunner;
 	private _toDispose: IDisposable[];
-	private _context: any;
 	private _currentDatabaseName: string;
 	private _isConnected: boolean;
 	private _databaseListDropdown: HTMLElement;
@@ -481,7 +569,6 @@ export class ListDatabasesActionItem implements IActionViewItem {
 	}
 
 	public setActionContext(context: any): void {
-		this._context = context;
 	}
 
 	public isEnabled(): boolean {
